+++ src/test/extra.rs
@@ -1,0 +1,2 @@
+//! Normally this is empty, but when extra tests are built, this is
+//! modified.  However those changes should not be checked in.